Change bar chart segment size to a float
As the height of a segment may not perfectly align to an int, using an
int means a loss of precision in the scale of the bars. This becomes
apparent as the maximum number of segments needed grows. Eventually,
this int would hit zero, and no bars would be drawn. Using a float
allows the bars to utilize all available space and avoid hitting a zero
segment height (up to a certain point. I have yet to determine when the
precision of a float is too small that the segment height would become
zero.).

+++ src/com/jgrocho/dicetracker/BarChartView.java
@@ -98,7 +98,7 @@ public class BarChartView extends View implements Rolls.OnChangeListener {
         // Draw bars
         int maxCount = max(mRolls.getRolls());
         if (maxCount > 0) {
-            float segmentHeight = mChartHeight / maxCount;
+            float segmentHeight = ((float) mChartHeight) / ((float) maxCount);
             for (int i = 0; i < 11; i++) {
                 if (mRolls.getAt(i) > 0) {
                     float left = i * mBarWidth;
